+++ dflib/src/main/java/org/dflib/zip/SequentialZip.java
@@ -1,0 +1,66 @@
+package org.dflib.zip;
+
+import org.dflib.ByteSource;
+import org.dflib.ByteSources;
+
+import java.io.IOException;
+import java.io.InputStream;
+import java.util.ArrayList;
+import java.util.List;
+import java.util.function.Predicate;
+import java.util.zip.ZipEntry;
+import java.util.zip.ZipInputStream;
+
+class SequentialZip extends Zip {
+
+    private final ByteSource source;
+
+    public SequentialZip(Predicate<ZipEntry> notHiddenFilter, ByteSource source) {
+        super(notHiddenFilter);
+        this.source = source;
+    }
+
+    @Override
+    public List<ZipEntry> list(boolean includeFolders) {
+        Predicate<ZipEntry> filter = entryFilter(includeFolders);
+        return source.processStream(in -> extractEntries(in, filter));
+    }
+
+    @Override
+    public Zip includeHidden() {
+        return new SequentialZip(null, source);
+    }
+
+    @Override
+    public ByteSource source(String zipPath) {
+
+        if (zipPath.endsWith("/")) {
+            throw new IllegalArgumentException("Zip entry is a directory and can not be used to create a ByteSource: " + zipPath);
+        }
+
+        return new ZipStreamSeekByteSource(source, zipPath);
+    }
+
+    @Override
+    public ByteSources sources() {
+        return new ZipStreamByteSources(source, entryFilter(false));
+    }
+
+    private List<ZipEntry> extractEntries(InputStream in, Predicate<ZipEntry> filter) {
+        ZipInputStream zin = new ZipInputStream(in);
+        List<ZipEntry> entries = new ArrayList<>();
+        ZipEntry entry;
+
+        try {
+            while ((entry = zin.getNextEntry()) != null) {
+                if (filter.test(entry)) {
+                    entries.add(entry);
+                }
+            }
+        } catch (IOException e) {
+            throw new RuntimeException("Error reading zip entries", e);
+        }
+
+        return entries;
+    }
+}