+++ dflib/src/test/java/org/dflib/zip/Zip_ListTest.java
@@ -15,6 +15,35 @@ public class Zip_ListTest {
     @MethodSource(value = "org.dflib.zip.TestZips#one")
     void list(Zip zip) {
         List<String> names = zip.list().stream().map(ZipEntry::getName).sorted().collect(Collectors.toList());
+        assertEquals(List.of("a/test2.txt", "a/test3.txt",  "b/c/test4.txt", "test.txt"), names);
+    }
+
+    @ParameterizedTest
+    @MethodSource(value = "org.dflib.zip.TestZips#one")
+    void list_IncludeFolders(Zip zip) {
+        List<String> names = zip.list(true).stream().map(ZipEntry::getName).sorted().collect(Collectors.toList());
         assertEquals(List.of("a/", "a/test2.txt", "a/test3.txt", "b/", "b/c/", "b/c/test4.txt", "test.txt"), names);
+    }
+
+    @ParameterizedTest
+    @MethodSource(value = "org.dflib.zip.TestZips#two")
+    void list_noHidden(Zip zip) {
+        List<String> names = zip.list().stream().map(ZipEntry::getName).sorted().collect(Collectors.toList());
+        assertEquals(List.of("test/f1.csv", "test/sub/f2.csv"), names);
+    }
+
+    @ParameterizedTest
+    @MethodSource(value = "org.dflib.zip.TestZips#two")
+    void list_includeHidden(Zip zip) {
+        List<String> names = zip.includeHidden()
+                .list()
+                .stream().map(ZipEntry::getName).sorted().collect(Collectors.toList());
+
+        assertEquals(List.of(
+                "__MACOSX/._test",
+                "__MACOSX/test/._.DS_Store",
+                "__MACOSX/test/sub/._f2.csv",
+                "test/.DS_Store", "test/f1.csv",
+                "test/sub/f2.csv"), names);
     }
 }