+++ dflib/src/test/java/org/dflib/zip/Zip_RandomAccessTest.java
@@ -12,23 +12,23 @@ public class Zip_RandomAccessTest {
 
     @Test
     void ofFile() throws URISyntaxException {
-        File zipFile = new File(getClass().getResource("test.zip").toURI());
+        File zipFile = new File(getClass().getResource("test1.zip").toURI());
         assertTrue(zipFile.exists());
 
         Zip zip = Zip.of(zipFile);
-        assertTrue(zip instanceof Zip.RandomAccessZip);
+        assertTrue(zip instanceof RandomAccessZip);
     }
 
     @Test
     void ofFileName() throws URISyntaxException {
-        File zipFile = new File(getClass().getResource("test.zip").toURI());
+        File zipFile = new File(getClass().getResource("test1.zip").toURI());
         Zip zip = Zip.of(zipFile.getAbsolutePath());
-        assertTrue(zip instanceof Zip.RandomAccessZip);
+        assertTrue(zip instanceof RandomAccessZip);
     }
 
     @Test
     void ofPath() throws URISyntaxException {
-        Zip zip = Zip.of(Path.of(getClass().getResource("test.zip").toURI()));
-        assertTrue(zip instanceof Zip.RandomAccessZip);
+        Zip zip = Zip.of(Path.of(getClass().getResource("test1.zip").toURI()));
+        assertTrue(zip instanceof RandomAccessZip);
     }
 }